Rectification et integration acceuil
Integration des cards avec tout les informations necessaires
Utilisation de template.jsp au lieu de acceuil.jsp

+++ src/main/java/itu/Compatibilite/PosteDetails.java
@@ -102,7 +102,7 @@ public class PosteDetails {
     public String getPosteDateInsertion() { return posteDateInsertion; }
     public void setPosteDateInsertion(String posteDateInsertion) { this.posteDateInsertion = posteDateInsertion; }
 
-    public double getPosteSalaire() { return posteSalaire; }
+    public int getPosteSalaire() { return (int)posteSalaire; }
     public void setPosteSalaire(double posteSalaire) { this.posteSalaire = posteSalaire; }
 
     public String getPosteTitre() { return posteTitre; }
